requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
bitarray==2.6.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crc16==0.1.1
crc32c==2.3.post0
cryptography==41.0.7
ed25519==1.5
exceptiongroup==1.2.0
fastapi==0.109.0
flower==2.0.1
frozenlist==1.4.1
graphql-core==3.2.3
graphql_query==1.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
humanize==4.9.0
idna==3.6
Jinja2==3.1.3
kombu==5.3.5
Mako==1.3.0
MarkupSafe==2.1.4
mnemonic==0.21
multidict==6.0.4
mysql-connector-python==8.3.0
pillow==10.2.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.1
PyMySQL==1.1.0
PyNaCl==1.5.0
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytonlib==0.0.53
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
starlette==0.35.1
supervisor==4.2.5
ton==0.26
ton-client-py==1.44.3.0
tonsdk==1.0.13
tornado==6.4
tvm-valuetypes==0.0.10
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.26.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
